+++ app/dto/ScheduleSpeakRequestDto.py
@@ -1,5 +1,3 @@
-from typing import List
-
 from pydantic import BaseModel
 
 
@@ -7,4 +5,3 @@ class ScheduleSpeakRequestDto(BaseModel):
     schedule_id: int
     schedule_voice_Url: str
     target_time: str  # "10:00:00" 형식
-
