+++ app/dto/ScheduleTTSRequestDto.py
@@ -5,6 +5,6 @@ from pydantic import BaseModel
 
 class ScheduleTTSRequestDto(BaseModel):
     voice_id: int
-    alias : str
+    alias: str
     schedule_id: List[int]
     schedule_text: List[str]
