+++ app/service/gpt.py
@@ -1,6 +1,8 @@
 import os
+
 from dotenv import load_dotenv
 from openai import OpenAI
+
 from app.utils import parsing_json
 
 load_dotenv()
