+++ app/utils/parsing_json.py
@@ -1,5 +1,5 @@
-import re
 import json
+import re
 
 
 def extract_json_from_content(content):
